refactor(prover): move inline table tests to tests/
Relocate the #[cfg(test)] mod tests blocks out of the table source files
in prover/src/tables/ into prover/src/tests/, alongside the existing
per-table test files.

- 7 new files: page/register/load/memw/memw_aligned/memw_register/
  keccak_rnd _tests.rs
- decode and trace_builder inline tests appended to their existing
  decode_tests.rs / trace_builder_tests.rs
- is_register_op and collect_bitwise_from_keccak bumped to pub(crate)
  so the relocated trace_builder tests can reach them

bitwise.rs has no #[test] functions (only a #[cfg(test)] test helper)
and is left unchanged; trace_builder's #[cfg(test)] impl helper methods
also stay in place.
