+++ prover/src/tables/decode.rs
@@ -422,81 +422,3 @@ fn build_decode_table(
 
     TraceTable::new_main(data, cols::NUM_COLUMNS, 1)
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use executor::elf::Segment;
-
-    #[test]
-    fn test_tables_from_elf_single_executable_segment() {
-        // ADDI x1, x0, 42  (opcode: 0x02a00093)
-        // ADDI x2, x1, 10  (opcode: 0x00a08113)
-        let elf = Elf {
-            entry_point: 0x1000,
-            data: vec![Segment {
-                base_addr: 0x1000,
-                values: vec![0x02a00093, 0x00a08113],
-                is_executable: true,
-            }],
-        };
-
-        let tables = tables_from_elf(&elf).unwrap();
-
-        // Check DECODE table
-        assert_eq!(tables.pc_to_row.len(), 3); // 2 instructions + CPU padding
-        assert!(tables.pc_to_row.contains_key(&0x1000));
-        assert!(tables.pc_to_row.contains_key(&0x1004));
-        assert!(
-            tables
-                .pc_to_row
-                .contains_key(&super::super::cpu::CPU_PADDING_PC)
-        );
-    }
-
-    #[test]
-    fn test_tables_from_elf_mixed_segments() {
-        // Executable segment with instructions
-        // Data segment with data (not included in DECODE)
-        let elf = Elf {
-            entry_point: 0x1000,
-            data: vec![
-                Segment {
-                    base_addr: 0x1000,
-                    values: vec![0x02a00093], // ADDI instruction
-                    is_executable: true,
-                },
-                Segment {
-                    base_addr: 0x2000,
-                    values: vec![0xDEADBEEF, 0xCAFEBABE], // Data
-                    is_executable: false,
-                },
-            ],
-        };
-
-        let tables = tables_from_elf(&elf).unwrap();
-
-        // DECODE: only executable segment (1 instruction + CPU padding)
-        assert_eq!(tables.pc_to_row.len(), 2);
-        assert!(tables.pc_to_row.contains_key(&0x1000));
-        assert!(!tables.pc_to_row.contains_key(&0x2000)); // Data not in decode
-    }
-
-    #[test]
-    fn test_tables_from_elf_empty() {
-        let elf = Elf {
-            entry_point: 0x1000,
-            data: vec![],
-        };
-
-        let tables = tables_from_elf(&elf).unwrap();
-
-        // DECODE: only CPU padding entry
-        assert_eq!(tables.pc_to_row.len(), 1);
-        assert!(
-            tables
-                .pc_to_row
-                .contains_key(&super::super::cpu::CPU_PADDING_PC)
-        );
-    }
-}
