+++ prover/src/tests/mod.rs
@@ -21,12 +21,26 @@ pub mod disk_spill_tests;
 #[cfg(test)]
 pub mod dvrm_tests;
 #[cfg(test)]
+pub mod keccak_rnd_tests;
+#[cfg(test)]
+pub mod load_tests;
+#[cfg(test)]
 pub mod lt_bus_tests;
 #[cfg(test)]
 pub mod lt_tests;
 #[cfg(test)]
+pub mod memw_aligned_tests;
+#[cfg(test)]
+pub mod memw_register_tests;
+#[cfg(test)]
+pub mod memw_tests;
+#[cfg(test)]
 pub mod mul_tests;
 #[cfg(test)]
+pub mod page_tests;
+#[cfg(test)]
 pub mod prove_elfs_tests;
+#[cfg(test)]
+pub mod register_tests;
 #[cfg(test)]
 pub mod trace_builder_tests;
